requirements/base.txt
django-extensions==1.5.6
ipython==4.0.0
python-dateutil==2.4.2
pytz
readline
